refactor: move auditLog config to camunda.data.audit-log
