+++ configuration/src/main/java/io/camunda/configuration/Data.java
@@ -22,6 +22,9 @@ public class Data {
   /** How often we take snapshots of streams (time unit) */
   private Duration snapshotPeriod = Duration.ofMinutes(5);
 
+  /** This section allows to configure the audit log setup. */
+  @NestedConfigurationProperty private AuditLog auditLog = new AuditLog();
+
   /** This section allows to configure primary Zeebe's data storage. */
   @NestedConfigurationProperty private PrimaryStorage primaryStorage = new PrimaryStorage();
 
@@ -33,6 +36,14 @@ public class Data {
 
   /** This section allows configuring exporters */
   private Map<String, Exporter> exporters = new HashMap<>();
+
+  public AuditLog getAuditLog() {
+    return auditLog;
+  }
+
+  public void setAuditLog(final AuditLog auditLog) {
+    this.auditLog = auditLog;
+  }
 
   public Duration getSnapshotPeriod() {
     return UnifiedConfigurationHelper.validateLegacyConfiguration(
